+++ extensions/redis/ActiveQuery.php
@@ -305,6 +305,7 @@ class ActiveQuery extends Component implements ActiveQueryInterface
      * If this parameter is not given, the `db` application component will be used.
      * @param string $type the type of the script to generate
      * @param string $columnName
+     * @throws NotSupportedException
      * @return array|bool|null|string
      */
     protected function executeScript($db, $type, $columnName = null)