+++ framework/base/Security.php
@@ -98,7 +98,7 @@ class Security extends Component
     }
 
     /**
-     * Encrypts data using a cyptograhic key.
+     * Encrypts data using a cryptograhic key.
      * Derives keys for encryption and authentication from the input key using HKDF and a random salt,
      * which is very fast relative to [[encryptByPassword()]]. The input key must be properly
      * random -- use [[generateRandomKey()]] to generate keys.
@@ -282,7 +282,7 @@ class Security extends Component
 
     /**
      * Derives a key from the given input key using the standard HKDF algorithm.
-     * Implements HKDF spcified in [RFC 5869](https://tools.ietf.org/html/rfc5869).
+     * Implements HKDF specified in [RFC 5869](https://tools.ietf.org/html/rfc5869).
      * Recommend use one of the SHA-2 hash algorithms: sha224, sha256, sha384 or sha512.
      * @param string $algo a hash algorithm supported by `hash_hmac()`, e.g. 'SHA-256'
      * @param string $inputKey the source key