+++ extensions/debug/CHANGELOG.md
@@ -1,8 +1,14 @@
 Yii Framework 2 debug extension Change Log
 ==========================================
 
-2.0.0 under development
+2.0.1 under development
 -----------------------
+
+- Enh #5600: Allow configuring debug panels in `yii\debug\Module::panels` as panel class name strings (qiangxue)
+
+
+2.0.0 October 12, 2014
+----------------------
 
 - no changes in this release.